test: mapped punk gateway tests
